Pulling Zipper hole shifting and axes into separate traits.
This removes the need in implicit conversion for Zipper axes support.
